optimize tests
